requirements.txt.txt
pygame
pywin32
matplotlib
python-vlc
requests